requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
DateTimeRange==1.0.0
Django==3.1.1
django-crispy-forms==1.9.2
idna==2.10
mbstrdecoder==1.0.0
Pillow==7.2.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
reportlab==3.5.50
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
typepy==1.1.1
urllib3==1.25.10
